requirements.txt
--index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio


attrs==24.3.0
backcall==0.2.0
bleach==6.1.0
cffi==1.16.0
cloudpickle==3.0.0
colorama==0.4.6
cycler==0.12.1
dask==2024.1.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
imageio==2.33.0
importlib-metadata==7.0.0
ipykernel==6.28.0
ipython==8.20.0
ipython-genutils
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.20.0
jupyter-client==8.6.0
jupyter-core==5.7.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mistune==3.0.2
nbconvert==7.12.0
nbformat==5.9.2
networkx==3.2
notebook==7.0.6
numpy==1.26.4
opencv-python
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
pycparser==2.21
pygit
Pygments==2.17.2
pyparsing==3.1.1
pyrsistent==0.20.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.5.0
pywin32==306
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==25.1.2
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
Send2Trash==1.8.2
six==1.16.0
terminado==0.18.0
testpath==0.6.0
toolz==0.12.0
torch
torchvision
tornado==6.4
tqdm==4.66.2
traitlets==5.14.0
wcwidth==0.2.13
webencodings
zipp==3.17.0